Support getFile API + move EndowmentPermissions
